feat: Allow custom todo/done files via command line
It may be necessary to track some commits in different files (e.g. if some are considered higher priority). This allows a custom results file to be specified for review.results and review.html.done, and a custom todo file for review.results.todo.

In support of this change, getopts was adopted. Thanks to <https://www.shellscript.sh/tips/getopts/> for the tips :-).

The help checks in review.init and review.park were changed so that passing /anything/ causes the help message to be shown and the script to exit. Also: slight readability improvements in review.next.

BREAKING CHANGE: The "--help" command-line option is no longer supported (due to the adoption of getopts).